requirements.txt
fastapi>=0.119.0
httpx>=0.28.1
python-dotenv>=1.1.1
uvicorn>=0.37.0
google-cloud-speech>=2.27.0
google-cloud-texttospeech>=2.17.0
python-multipart>=0.0.9
pyaudio>=0.2.14
pynput>=1.7.6
